[COMMIT] PUSH

+++ app/__init__.py
@@ -106,16 +106,17 @@ def show_one_thing(id):
 def add_a_thing():
     # Get the data from the form
     name  = request.form.get("name")
-    price = request.form.get("price")
 
     # Sanitise the inputs
     name = html.escape(name)
-    price = html.escape(price)
+
+    # Get the user id from the session
+    user_id = session["user_id"]
 
     with connect_db() as client:
         # Add the thing to the DB
-        sql = "INSERT INTO things (name, price) VALUES (?, ?)"
-        values = [name, price]
+        sql = "INSERT INTO things (name, user_id) VALUES (?, ?)"
+        values = [name, user_id]
         client.execute(sql, values)
 
         # Go back to the home page
@@ -147,7 +148,7 @@ def add_a_user():
         client.execute(sql, values)
 
         # Go back to the home page
-        flash(f"Thing '{name}' added", "success")
+        flash(f"User '{name}' added", "success")
         return redirect("/")
     
 #-----------------------------------------------------------
@@ -184,15 +185,15 @@ def login_user():
 
         # Else
         flash("Incorrect login credentials", "error")
-        redirect("/login")
+        return redirect("/login")
 
 
-
-        client.execute(sql, values)
-
-        # Go back to the home page
-        flash(f"Thing added", "success")
-        return redirect("/")
+@app.get("/logout")
+def logout():
+    session.pop("user_id")
+    session.pop("user_name")
+    flash("You have been logged out")
+    return redirect ("/")
 
 
 #-----------------------------------------------------------
@@ -201,9 +202,9 @@ def login_user():
 @app.get("/delete/<int:id>")
 def delete_a_thing(id):
     with connect_db() as client:
-        # Delete the thing from the DB
-        sql = "DELETE FROM things WHERE id=?"
-        values = [id]
+        # Delete the thing from the DB, checking vvvvvvvvv that we are the owner
+        sql = "DELETE FROM things WHERE id=? AND user_id =?"
+        values = [id, session.user_id]
         client.execute(sql, values)
 
         # Go back to the home page
